Fix 32-bit overflow in raw bytes counter
git-svn-id: svn+ssh://svn.code.sf.net/p/turbovnc/code/trunk@1858 799e4f7b-5fd2-41f6-823c-2ecc41bc7f0b

+++ java/VncCanvas.java
@@ -67,8 +67,8 @@ class VncCanvas extends Canvas
   int statNumRectsHextile;      // Hextile-encoded rectangles
   int statNumRectsRaw;          // Raw-encoded rectangles
   int statNumRectsCopy;         // CopyRect rectangles
-  int statNumBytesEncoded;      // number of bytes in updates, as received
-  int statNumBytesDecoded;      // number of bytes, as if Raw encoding was used
+  long statNumBytesEncoded;     // number of bytes in updates, as received
+  long statNumBytesDecoded;     // number of bytes, as if Raw encoding was used
 
   // ZRLE encoder's data.
   byte[] zrleBuf;
